requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
prisma
python-jose[cryptography]
redis